requirements.txt
python-dateutil>=2.8.2
tzlocal>=3.0
argparse>=1.4.0
psutil>=5.9.0
pytz>=2022.5
PyYAML>=6.0
freezegun>=1.5.0